test: Run all tests in parallel
Run all tests configured tests in parallel using goroutines. We wait
until all tests completed and fail the test if one of the tests failed.

We don't limit the number of parallel tests. For ramenctl we never
expect large list of tests so this should not be an issue. For ramen we
have now 6 possible tests, and we will likely double or tripple the
number of tests in the future. We may want to limit the number of
parallel tests later.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>