report: Fix yaml stale state serialization
The stale emoji was serialized to:

    state: "stale \U0001F7E0"

Instead of:

    state: stale 🟠

Which is not helpful.

Replace with an emoji that serialized well:

    state: stale ⭕

I think the issue is unicode values outside of the basic plane
(0000-FFFF). The previous emoji was U+1F7E0, the new emoji is U+2B55.

Add tests to make it easier to inspect the yaml output and ensure that
the values are marshaled and unmarshaled correctly.

Fixes: #285
Signed-off-by: Nir Soffer <nsoffer@redhat.com>