validate: Fix secrets in testdata
It seems that the secrets "sanitized" before adding them to the repo by
replacing the data fields with "FAKESECRET". This change is invalid
since secret data fields must be base64 encoded. Unmarshalling the
secrets fail with:

    error unmarshaling JSON: while decoding JSON: illegal base64 data at input byte 12

Fixed by modifying the fields to valida base64 value which is not a real
secret.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>