validate: Validate ramen deployments
Validate ramen deployments on the hub and managed clusters.

Deployments are generic have nothing specific to ramen or to the
cluster, so we have a generic validateDeployment() function. We call the
helper with the ramen deployment name and namespace.

Deployment is not part of the core k8s package we have a new apps.go
file to handling deployments. This includes also
validatedDeploymentCondition() since deployments do not use the standard
conditions.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>